requirements.txt
google-cloud-speech==2.33.0
pyaudio==0.2.14 